Refactor common code to abstract Disks class
